# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-signature documentation",
        "description": "This is the API for E-Signature app",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/documents": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "Get documents list",
                "operationId": "3d2cba87c070891669e4a985a0765a66",
                "responses": {
                    "200": {
                        "description": "Documents list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/DocumentResource"
                                                    }
                                                },
                                                "meta": {
                                                    "$ref": "#/components/schemas/Pagination"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/documents/{documentId}": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "Get document",
                "operationId": "f80dc041dbcb6a69290d5d87dc8217c7",
                "parameters": [
                    {
                        "name": "documentId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1,
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Document info",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "object",
                                                    "allOf": [
                                                        {
                                                            "$ref": "#/components/schemas/DocumentResource"
                                                        }
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/documents/upload": {
            "post": {
                "tags": [
                    "Documents"
                ],
                "summary": "Upload a PDF file",
                "operationId": "70c73d55e38e07ba8e12191e5a9b32af",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Document uploaded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocumentResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Invalid file format"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/documents/{documentId}/send": {
            "post": {
                "tags": [
                    "Documents"
                ],
                "summary": "Send document for signature",
                "operationId": "8f30a9a25e752010ce24b50626e8f195",
                "parameters": [
                    {
                        "name": "documentId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1,
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 12
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Document successfully sent for signature",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SignatureResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/signatures/{signatureRequestId}/sign": {
            "post": {
                "tags": [
                    "Signatures"
                ],
                "summary": "Sign document",
                "operationId": "985c0775fb798ca2327c6654d60455b7",
                "parameters": [
                    {
                        "name": "signatureRequestId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1,
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "sign_text"
                                ],
                                "properties": {
                                    "sign_text": {
                                        "type": "string",
                                        "example": "My sign!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Document successfully signed"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Something went wrong."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/signatures/requests": {
            "get": {
                "tags": [
                    "Signatures"
                ],
                "summary": "Get signature requests list",
                "operationId": "c927667d8172b2dfdff21450ba42a51c",
                "responses": {
                    "200": {
                        "description": "Signature requests list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/SignatureResource"
                                                    }
                                                },
                                                "meta": {
                                                    "$ref": "#/components/schemas/Pagination"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "DocumentResource": {
                "title": "Document Resource",
                "description": "Document data resource",
                "properties": {
                    "id": {
                        "description": "Transform the resource into an array.",
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "name": {
                        "type": "string",
                        "example": "rent_contract"
                    },
                    "path": {
                        "type": "string",
                        "example": "/tmp/pdf/test.pdf"
                    }
                },
                "type": "object"
            },
            "SignatureResource": {
                "title": "Signature Resource",
                "description": "Signature data resource",
                "properties": {
                    "id": {
                        "description": "Transform the resource into an array.",
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "document_id": {
                        "type": "integer",
                        "example": 27
                    },
                    "status": {
                        "type": "string",
                        "example": "pending"
                    }
                },
                "type": "object"
            },
...
            "Pagination": {
                "description": "Pagination object",
                "required": [
                    "pagination"
                ],
                "properties": {
                    "pagination": {
                        "type": "object",
                        "nullable": false,
                        "allOf": [
                            {
                                "required": [
                                    "count",
                                    "current_page",
                                    "per_page",
                                    "total",
                                    "total_pages",
                                    "links"
                                ],
                                "properties": {
                                    "count": {
                                        "description": "Count result on page",
                                        "type": "integer",
                                        "nullable": false
                                    },
                                    "current_page": {
                                        "description": "Current page number",
                                        "type": "integer",
                                        "nullable": false
                                    },
                                    "per_page": {
                                        "description": "Curent page limit",
                                        "type": "integer",
                                        "nullable": false
                                    },
                                    "total": {
                                        "description": "Total results",
                                        "type": "integer",
                                        "nullable": false
                                    },
                                    "total_pages": {
                                        "description": "Total pages",
                                        "type": "integer",
                                        "nullable": false
                                    },
                                    "links": {
                                        "description": "Links for pagination",
                                        "type": "object",
                                        "nullable": false,
                                        "allOf": [
                                            {
                                                "properties": {
                                                    "next": {
                                                        "description": "Link for next page",
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "previous": {
                                                        "description": "Link for previous page",
                                                        "type": "string",
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        ]
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}